requirement.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.2
reportlab==4.0.4
pandas==2.1.3
plotly==5.18.0
streamlit==1.28.1
python-dotenv==1.0.0
typing-extensions==4.8.0
requests==2.31.0
numpy==1.24.3